remove BitTypeDescriptor
