+++ hibernate-core/src/main/java/org/hibernate/dialect/SybaseAnywhereDialect.java
@@ -18,8 +18,6 @@ import org.hibernate.sql.ast.SqlAstTranslatorFactory;
 import org.hibernate.sql.ast.spi.StandardSqlAstTranslatorFactory;
 import org.hibernate.sql.ast.tree.Statement;
 import org.hibernate.sql.exec.spi.JdbcOperation;
-import org.hibernate.type.descriptor.sql.BitTypeDescriptor;
-import org.hibernate.type.descriptor.sql.SqlTypeDescriptor;
 
 import java.sql.Types;
 
@@ -76,13 +74,6 @@ public class SybaseAnywhereDialect extends SybaseDialect {
 	@Override
 	public boolean supportsTimezoneTypes() {
 		return true;
-	}
-
-	@Override
-	protected SqlTypeDescriptor getSqlTypeDescriptorOverride(int sqlCode) {
-		return sqlCode == Types.BOOLEAN
-				? BitTypeDescriptor.INSTANCE
-				: super.getSqlTypeDescriptorOverride( sqlCode );
 	}
 
 	@Override
